XML_DataReadingMaskingWriting
The program, which reads data from an XML file (books.xml) and save it
as a separate file (new.xml) masking credit card information (number and
CVC) according to the following scheme: Number: 412 827 ****** 9009 -
Visiblemay be only the last four and the first six digits. CVC: ***. The
algorithm includes situations where the card number is the wrong length,
empty or will contain invalid characters.